feat: SingletonComponent에서 싱글톤으로 GamsungDatabase, MenuDao 관리
