+++ app/src/main/java/com/erica/gamsung/core/presentation/MainActivity.kt
@@ -14,7 +14,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.erica.gamsung.core.di.GamsungDatabase
+import com.erica.gamsung.core.data.GamsungDatabase
 import com.erica.gamsung.core.presentation.theme.GamsungTheme
 import com.erica.gamsung.menu.data.remote.MenuApi
 import com.erica.gamsung.menu.data.repository.MenuRepositoryImpl
@@ -24,13 +24,17 @@ import com.erica.gamsung.store.presentation.InputStoreScreen
 import com.erica.gamsung.uploadTime.presentation.CalendarViewModel
 import com.erica.gamsung.uploadTime.presentation.MyCalendarScreen
 import com.erica.gamsung.uploadTime.presentation.MyScheduleScreen
+import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
 
 // 동일한 viewmodel을 2개의 page가 공유하기 위해서는 hilt를 이용한 DI가 필요하다고 한다.
 // 일단은 상위 컴포넌트에서 생성해서 사용.
+@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
+    @Inject lateinit var database: GamsungDatabase
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val database = GamsungDatabase.getInstance(this)!!
         setContent {
             GamsungTheme {
                 Surface(
